cartera/requirements.txt
Django==5.0.7
python-dotenv==1.0.1
psycopg[binary]==3.1.19
django-filter==24.2
